코드 리팩토링 : Router 코드 리팩토링 기능별로 리팩토링
